refactor all posts/server code to just posts
