+++ lib/content/posts/rendering.ts
@@ -1,17 +1,13 @@
-import { useMemo, createElement, Fragment } from 'react'
+import { createElement, Fragment } from 'react'
 import rehypeReact from 'rehype-react'
 import { ComponentsWithoutNodeOptions } from 'rehype-react/lib/complex-types'
 import { unified } from 'unified'
 import { HastTree } from './types'
 
-export const useReactFromHast = (
+export const convertToReact = (
   content: HastTree,
   components: ComponentsWithoutNodeOptions['components']
 ): React.ReactNode =>
-  useMemo(
-    () =>
-      unified()
-        .use(rehypeReact, { createElement, Fragment, components })
-        .stringify(content),
-    [content, components]
-  )
+  unified()
+    .use(rehypeReact, { createElement, Fragment, components })
+    .stringify(content)
