+++ lib/content/posts/validation.ts
