+++ lib/pages/posts/[slug].ts
@@ -1,7 +1,7 @@
-import { convertRawStringToPost, getPost } from 'lib/content/posts/server'
+import { convertRawStringToPost, getPost } from 'lib/content/posts'
 import { HastTree, Post, PostMeta } from 'lib/content/posts/types'
-import { imageService } from 'lib/content/posts/server/imageService'
-import { parseToHast } from 'lib/content/posts/server/parsing'
+import { imageService } from 'lib/content/posts/imageService'
+import { parseToHast } from 'lib/content/posts/parsing'
 
 import { A11yStaticImageData } from 'lib/content/images'
 import { formatDate } from 'lib/pages/posts/utils'
